BRS/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
entrypoints==0.3
graphviz==0.17
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
lightfm==1.16
lightfm-dataset-helper==0.1.2
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.4
Pillow==8.4.0
plotly==5.3.1
pytz==2021.3
requests==2.26.0
scikit-learn==1.0
scikit-surprise==1.1.1
scipy==1.7.1
tenacity==8.0.1
terminado==0.9.4
threadpoolctl==3.0.0
urllib3==1.26.7
webencodings==0.5.1
